feat: Add dependency injection container setup
This commit adds a new file `container.dart` that contains the setup for dependency injection using the `get_it` package. It registers lazy singletons for `SharedPreferences`, `FlutterSecureStorage`, and `Dio` instances. This setup enhances the modularity and testability of our codebase by providing a centralized location for managing dependencies. It also improves the overall performance and maintainability of our application.

Related to recent user commits:
- feat: Add `flutter_screenutil` and `flutter_secure_storage` dependencies
- Added `ApiResult` Freezed Models
- Introduced ApiResult and RepoResult types
- Added JSON serialization for API response models
- Added API response models
- feat: Add TryCallApi class for handling API calls with error handling
- chore: Add ApiErrorHandler class for handling Dio exceptions
- Add Dio networking setup with interceptors
- Added development and production flavors
- Updated dependencies for enhanced functionality