Refactor app initialization and DI setup
Streamlined the main app initialization process and integrated a new dependency injection system to enhance modularity and ease future development efforts. This change involved removing the boilerplate Flutter demo code and replacing it with a more scalable structure that leverages `ScreenUtil` for responsive UI design, introduces a centralized API constants management, and sets up a dependency injection framework right at the app's entry point. This refactoring is pivotal for supporting a wider range of device screens and simplifying the connectivity with backend services, thereby laying a solid foundation for the app's scalability and maintenance.