feat: Add extension for showing AlertDialog
This commit adds a new extension `AlertDialogExtension` on `BuildContext` to provide a convenient way to show an `AlertDialog` with customizable options. The extension includes methods for displaying a dialog box with a title, body, content widget, confirm and cancel buttons, and custom callbacks. Additionally, it introduces a `showPopUp` method for showing a pop-up with a specified content widget. This extension enhances the reusability and readability of code that requires showing alert dialogs or pop-ups, improving the overall user experience.

Related to recent user commits:
- Integrate router and ScreenUtil support
- feat: Add FlutterSecureStorage plugin registration for Linux and Windows
- feat: Add ApiConstants class for managing API base URLs
- feat: Refactor main app structure for production
- feat: Add `DocQuizAPP` class for initializing the Flutter application
- feat: Add dependency injection container setup
- feat: Add `flutter_screenutil` and `flutter_secure_storage` dependencies
- Added `ApiResult` Freezed Models
- Introduced ApiResult and RepoResult types
- Added JSON serialization for API response models
- Added API response models
- feat: Add TryCallApi class for handling API calls with error handling
- chore: Add ApiErrorHandler class for handling Dio exceptions
- Add Dio networking setup with interceptors
- Added development and production flavors
- Updated dependencies for enhanced functionality