feat: Add extension for showing BottomSheet
This commit adds a new extension `BottomSheet` on `BuildContext` to provide a convenient way to show a bottom sheet with customizable options. The extension includes a method for displaying a bottom sheet with a child widget. This extension enhances the reusability and readability of code that requires showing bottom sheets, improving the overall user experience.