feat: Add LoginCubit and LoginState classes
This commit adds the `LoginCubit` and `LoginState` classes to the project. The `LoginCubit` is a `Cubit` subclass that manages the state of the login feature, while the `LoginState` is a freezed union class that represents the different states of the login feature. This addition lays the foundation for implementing the login functionality in the app.